[Controller] making featues around external wrench estimation common one
