+++ robots/dragon/include/dragon/control/full_vectoring_control.h
@@ -52,11 +52,7 @@ namespace aerial_robot_control
   {
   public:
     DragonFullVectoringController();
-    ~DragonFullVectoringController()
-    {
-      wrench_estimate_thread_.interrupt();
-      wrench_estimate_thread_.join();
-    }
+    ~DragonFullVectoringController() = default;
 
     void initialize(ros::NodeHandle nh, ros::NodeHandle nhp,
                     boost::shared_ptr<aerial_robot_model::RobotModel> robot_model,
@@ -69,7 +65,6 @@ namespace aerial_robot_control
     ros::Publisher flight_cmd_pub_; //for spinal
     ros::Publisher gimbal_control_pub_;
     ros::Publisher target_vectoring_force_pub_;
-    ros::Publisher estimate_external_wrench_pub_;
     ros::Publisher rotor_interfere_wrench_pub_;
     ros::Publisher interfrence_marker_pub_;
 
@@ -82,16 +77,6 @@ namespace aerial_robot_control
     bool gimbal_vectoring_check_flag_;
     double allocation_refine_threshold_;
     int allocation_refine_max_iteration_;
-    Eigen::VectorXd target_wrench_acc_cog_;
-
-    /* external wrench */
-    std::mutex wrench_mutex_;
-    boost::thread wrench_estimate_thread_;
-    Eigen::VectorXd init_sum_momentum_;
-    Eigen::VectorXd est_external_wrench_;
-    Eigen::MatrixXd momentum_observer_matrix_;
-    Eigen::VectorXd integrate_term_;
-    double prev_est_wrench_timestamp_;
 
     bool rotor_interfere_compensate_;
     double fz_bias_;
@@ -115,7 +100,6 @@ namespace aerial_robot_control
     double overlap_dist_inter_joint_thresh_;
 
 
-    void externalWrenchEstimate();
     const Eigen::VectorXd getTargetWrenchAccCog()
     {
       std::lock_guard<std::mutex> lock(wrench_mutex_);
